+++ lib/td_dd_web/controllers/data_structure_version_controller.ex
@@ -8,6 +8,8 @@ defmodule TdDdWeb.DataStructureVersionController do
   alias Ecto
   alias TdCache.TemplateCache
   alias TdDd.DataStructures
+  alias TdDd.DataStructures.DataStructure
+  alias TdDd.DataStructures.DataStructureVersion
   alias TdDd.DataStructures.Tags
   alias TdDdWeb.SwaggerDefinitions
 
@@ -28,6 +30,7 @@ defmodule TdDdWeb.DataStructureVersionController do
     :domain,
     :external_id,
     :links,
+    :with_metadata_protected,
     :metadata_versions,
     :parents,
     :profile,
@@ -78,10 +81,20 @@ defmodule TdDdWeb.DataStructureVersionController do
 
   defp enrich_opts(%{user_id: user_id} = claims, data_structure) do
     Enum.filter(@enrich_attrs, fn
-      :profile -> can?(claims, view_data_structures_profile(data_structure))
-      :with_confidential -> can?(claims, manage_confidential_structures(data_structure))
-      :grants -> can?(claims, view_grants(data_structure))
-      _ -> true
+      :profile ->
+        can?(claims, view_data_structures_profile(data_structure))
+
+      :with_confidential ->
+        can?(claims, manage_confidential_structures(data_structure))
+
+      :grants ->
+        can?(claims, view_grants(data_structure))
+
+      :with_metadata_protected ->
+        can?(claims, view_protected_metadata([DataStructure, DataStructureVersion]))
+
+      _ ->
+        true
     end) ++ [user_id: user_id]
   end
 
@@ -138,7 +151,10 @@ defmodule TdDdWeb.DataStructureVersionController do
   end
 
   defp get_data_structure_version(data_structure_id, "latest", opts) do
-    DataStructures.get_latest_version(data_structure_id, opts)
+    DataStructures.get_latest_version(
+      data_structure_id,
+      opts
+    )
   end
 
   defp get_data_structure_version(data_structure_id, version, opts) do